+++ CHANGES.md
@@ -7,7 +7,8 @@ Table of new and noteworthy changes:
 | 0.1.19 | __New Release__                                                                                                           |
 | 0.1.19 | First beginnings of a GTK+ based GUI for Eruption                                                                         |
 | 0.1.19 | Add support for USB HID based event handling for mouse devices. This should enable the extra buttons on supported devices |
-| 0.1.19 | Update Lua scripts to handle the unified canvas; add script `batique-mouse.lua` specifically for mice                     |
+| 0.1.19 | Update Lua scripts to better handle the unified canvas                                                                    |
+| 0.1.19 | Add profile `batique-mouse.profile` specifically for mice                                                                 |
 | 0.1.19 | Fixed brightnes bug: Wrong initial brightness, was only correct after a key has been pressed                              |
 | 0.1.19 | New profiles: "Red FX", "Red Wave", "Heartbeat: System Monitor", "Fireplace", "Flight (Perlin)"                           |
 | 0.1.19 | Improve stability of the core daemon as well as the process-monitor daemon                                                |